# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: goimg-datalayer API
  version: 2.0.0
  description: |
    A self-hosted image gallery backend inspired by Flickr and Chevereto.

    Features:
    - User management with JWT authentication
    - Image upload with automatic variant generation
    - Albums, tags, and social features (likes, comments)
    - Content moderation and abuse reporting
    - RESTful API with RFC 7807 error responses

    **Phase 2 Complete**: All Phase 2 features (Sprints 10-15) implemented.
    **Phase 3 Progress**: Sprints 16-20 complete (oEmbed, nested albums, custom variants, test coverage, trending tags, featured picks, groups/communities).
    **Current Sprint**: Sprint 23 in progress (test coverage improvement, documentation updates).
    Future features in backlog: video support (Sprint 21) and account tiers (Sprint 22).

  contact:
    name: goimg-datalayer API Team
    url: https://github.com/yegamble/goimg-datalayer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.goimg.com/v1
    description: Production server (example)
security:
  - bearerAuth: []
...
paths:
...
  /images/{id}:
    get:
      tags: [images]
      summary: Get image details
      description: Retrieve detailed information about a specific image
      operationId: getImage
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '200':
          description: Image retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          description: Image not found or not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    put:
      tags: [images]
      summary: Update image metadata
      description: |
        Update image metadata. Only the image owner can update their images.
        Admins can update any image.
      operationId: updateImage
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 255
                description:
                  type: string
                  maxLength: 2000
                visibility:
                  type: string
                  enum: [public, private, unlisted]
                tags:
                  type: array
                  items:
                    type: string
                  maxItems: 20
                  example: [sunset, mountains, nature]
      responses:
        '200':
          description: Image updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - cannot update another user's image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags: [images]
      summary: Delete image
      description: |
        Delete an image permanently. Only the image owner can delete their images.
        Moderators and admins can delete any image.
      operationId: deleteImage
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '204':
          description: Image deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - cannot delete another user's image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /images/{id}/variants/{size}:
    get:
      tags: [images]
      summary: Get image variant
      description: |
        Retrieve a specific image variant (resized version).

        Available sizes:
        - thumbnail: 160px max width (WebP, quality 82)
        - small: 320px max width (WebP, quality 85)
        - medium: 800px max width (WebP, quality 85)
        - large: 1600px max width (WebP, quality 88)
        - original: unchanged dimensions (original format, quality 100 - maximum quality with security re-encoding)

        Aspect ratio is preserved (contain, not crop).
      operationId: getImageVariant
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
        - name: size
          in: path
          required: true
          description: Variant size
          schema:
            type: string
            enum: [thumbnail, small, medium, large, original]
      responses:
        '200':
          description: Image variant retrieved successfully
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Image or variant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /images/{id}/qr:
    get:
      tags: [images]
      summary: Get image share QR code
      description: |
        Generate a PNG QR code for sharing a public image.
        The QR code points to the SEO/social preview page:
        `/images/{id}/preview`.

        Optional authentication is supported, but only public images
        are shareable via this endpoint.
      operationId: getImageQRCode
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
        - name: size
          in: query
          required: false
          description: QR code size in pixels
          schema:
            type: integer
            minimum: 128
            maximum: 1024
            default: 256
      responses:
        '200':
          description: QR code generated successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid image ID or size parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found or not publicly shareable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: QR code generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication. Include the access token in the Authorization header:

        ```
        Authorization: Bearer <access_token>
        ```

        Access tokens expire after 15 minutes. Use the refresh endpoint to obtain a new token.
  parameters:
...
    ImageIdParam:
      name: id
      in: path
      required: true
      description: Image ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    Image:
      type: object
      required: [id, owner_id, owner, visibility, mime_type, width, height, variants, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        owner_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        owner:
          type: object
          required: [id, username]
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            display_name:
              type: string
              nullable: true
          example:
            id: 123e4567-e89b-12d3-a456-426614174001
            username: johndoe
            display_name: John Doe
        title:
          type: string
          nullable: true
          example: Sunset over mountains
        description:
          type: string
          nullable: true
          example: Beautiful sunset captured in Rocky Mountains
        visibility:
          type: string
          enum: [public, private, unlisted]
          example: public
        mime_type:
          type: string
          example: image/jpeg
        file_size:
          type: integer
          description: File size in bytes
          example: 2048576
        width:
          type: integer
          description: Image width in pixels
          example: 4000
        height:
          type: integer
          description: Image height in pixels
          example: 3000
        variants:
          type: object
          properties:
            thumbnail:
              $ref: '#/components/schemas/ImageVariant'
            small:
              $ref: '#/components/schemas/ImageVariant'
            medium:
              $ref: '#/components/schemas/ImageVariant'
            large:
              $ref: '#/components/schemas/ImageVariant'
            original:
              $ref: '#/components/schemas/ImageVariant'
        tags:
          type: array
          items:
            type: string
          example: [sunset, mountains, nature, landscape]
        view_count:
          type: integer
          example: 150
        like_count:
          type: integer
          example: 42
        comment_count:
          type: integer
          example: 8
        created_at:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2024-01-15T11:00:00Z
    ImageVariant:
      type: object
      required: [url, width, height]
      properties:
        url:
          type: string
          format: uri
          example: https://cdn.goimg.com/images/123e4567/medium.jpg
        width:
          type: integer
          example: 800
        height:
          type: integer
          example: 600

    # ==================== IPFS Schemas ====================
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: https://api.goimg.com/problems/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: Request validation failed
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: /api/v1/images
        traceId:
          type: string
          description: A unique identifier for tracing the request
          example: abc123-def456
        errors:
          type: array
          description: Detailed validation errors (for 400 responses)
          items:
            type: object
            properties:
              field:
                type: string
                example: title
              message:
                type: string
                example: must not exceed 255 characters

    # ==================== Session ====================
...